requirements/common.txt
Django==1.9.1
psycopg2==2.6.1
djangorestframework==3.2.3
requests==2.9.1
celery[redis]==3.1.23
django-celery==3.1.17
